Add modified field to Game
Previously, on updating an entry the created date would be changed. This
had the effect of pushing recently updated entries to the top of the
gallery page and allowed folks to keep their game listed first by updating
their entry. While I'd hope this would be a non-issue, it is probably best
to guard against it.